+++ crates/python/src/clustering/label_propagation.rs
@@ -46,7 +46,7 @@ impl PyLabelPropagation {
             GraphType::DiGraph(graph) => self.instance.detect_communities(graph),
         };
         map.into_iter()
-            .map(|(u, c)| (u.index() as usize, c))
+            .map(|(u, c)| (u.index(), c))
             .collect::<HashMap<_, _>>()
     }
 }
