+++ crates/python/src/clustering/louvain.rs
@@ -47,7 +47,7 @@ impl PyLouvain {
             GraphType::DiGraph(graph) => self.instance.detect_communities(graph),
         };
         map.into_iter()
-            .map(|(u, c)| (u.index() as usize, c))
+            .map(|(u, c)| (u.index(), c))
             .collect::<HashMap<_, _>>()
     }
 }
